requirements.txt
Django==3.2.10
environ==1.0
Pillow==9.3.0
mysqlclient